final commit before quals...
removed aim to substation code because it was buggy and behavior was too unpredictable.  worked well when cycling solely to the tall junction but never anywhere else.

added an "auto junction offset" that can be manually set by the driver(s) before autonomous init to account for junction leaning.  It's not perfect, but it's better than nothing.

switched to lines in auto because it didn't make sense to use splines, lines are much more predictable and the robot isn't following a complex path

fixed some small bugs